Fast Food Menu
In this assignment you create a project where you purchase burgers, fries, and sodas. You display each item's price and let the user to enter quantities for each of these three items. Your enter certain amount of cash to cover this purchase to cashier. Then press on Calculate button to calculate the change amount and display it in listed denominations of bills and coins (you have completed the coins change in assignment 2). When you click on Clear All button it clears all the entries in textboxes and all calculated items in labels.  When you click on Exit button it closes the form (interface). See below screenshot for the completed project.  Read instructions below carefully.